src/collection/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
flask-marshmallow
marshmallow-sqlalchemy
pymysql
requests
mysql-connector-python
PyJWT
cryptography